refactor(repo C12): extract LLM cache to engine/repository/llm-cache.ts
Moves getCacheKey, getCachedResult, setCachedResult, clearCache,
deleteLLMCache into engine/repository/llm-cache.ts. Self-contained:
only depends on Database and createHash. No module state.

engine/repository.ts: 4772 → 4733 lines.
Build: clean. Tests: 970/970 pass.
Regression bench: 91.51% (baseline: 91.74%, well within ±5% gate).

Spec: docs/superpowers/specs/2026-05-20-kindx-strategic-refactor-program-design.md §5 cluster C12
