+++ engine/repository.ts
@@ -156,6 +156,7 @@ export type ExpandedQuery = {
 // All path symbols now re-exported via the barrel at the top of this file.
 // Imported back here for internal use:
 import { resolve, homedir, getPwd, getRealPath, getDefaultDbPath } from "./repository/paths.js";
+import { getCacheKey, getCachedResult, setCachedResult, clearCache, deleteLLMCache } from "./repository/llm-cache.js";
 
 // =============================================================================
 // Virtual Path Utilities (kindx://)
@@ -909,52 +910,12 @@ export function getIndexHealth(db: Database): IndexHealthInfo {
 }
 
 // =============================================================================
-// Caching
+// Caching / LLM cache — moved to engine/repository/llm-cache.ts (W1 C12)
 // =============================================================================
-
-export function getCacheKey(url: string, body: object): string {
-  const hash = createHash("sha256");
-  hash.update(url);
-  hash.update(JSON.stringify(body));
-  return hash.digest("hex");
-}
-
-export function getCachedResult(db: Database, cacheKey: string): string | null {
-  const row = db.prepare(`SELECT result FROM llm_cache WHERE hash = ?`).get(cacheKey) as { result: string } | null;
-  return row?.result || null;
-}
-
-export function setCachedResult(db: Database, cacheKey: string, result: string): void {
-  const now = new Date().toISOString();
-  db.prepare(`INSERT OR REPLACE INTO llm_cache (hash, result, created_at) VALUES (?, ?, ?)`).run(cacheKey, result, now);
-  // Tier-1 perf: high-water-mark eviction. Only run the O(n log n) DELETE
-  // when the table has actually grown past the high-water threshold; avoids
-  // the random-sampling spike where the same DELETE could fire 100 times in
-  // quick succession by coincidence on a hot path.
-  if (Math.random() < 0.01) {
-    const row = db.prepare(`SELECT COUNT(*) AS c FROM llm_cache`).get() as { c: number };
-    if (row.c > 1500) {
-      db.exec(`DELETE FROM llm_cache WHERE hash NOT IN (SELECT hash FROM llm_cache ORDER BY created_at DESC LIMIT 1000)`);
-    }
-  }
-}
-
-export function clearCache(db: Database): void {
-  db.exec(`DELETE FROM llm_cache`);
-}
 
 // =============================================================================
 // Cleanup and maintenance operations
 // =============================================================================
-
-/**
- * Delete cached LLM API responses.
- * Returns the number of cached responses deleted.
- */
-export function deleteLLMCache(db: Database): number {
-  const result = db.prepare(`DELETE FROM llm_cache`).run();
-  return result.changes;
-}
 
 /**
  * Remove inactive document records (active = 0).
